chore: simplify today's changes (comments, dedup, mutation safety)
- App.svelte: tighten authorship effect — drop module-level _lastAuthorshipVisible dual-purpose flag in favor of a single lastDispatchedAuthorship sentinel; trim 14-line narrative comment to one line.
- annotation.ts: trim post-sync settle narrative; guard the 500ms setTimeout with rebuiltSinceMount so it no-ops when the observer already ran during the settling window.
- SettingsPopover.svelte: drop no-op `if (x !== null) x = null` guards on the activeSection effect.
- FormattingToolbar.svelte / file-opener.ts / fileUpload.ts / api-paths.ts / types.ts: trim narrative + issue-reference comments per project guidance (keep WHY, not WHAT/PR refs).
- file-opener.ts: replace `Map.forEach` while deleting with `for (const k of [...map.keys()])` snapshot iteration (Y.Map iteration under concurrent mutation is undefined).
- tests/helpers/docx-fixtures.ts: extract `buildDocxWithComments` from two duplicated copies in file-opener-{transact-batching,cleanup-on-failure}.test.ts.
- tests/helpers/fetch-source-asserts.ts: extract `expectFetchWithTimeoutEndpoint` + `countMatches` from two duplicated copies in tests/channel/run-timeouts.test.ts and tests/monitor/runtime-fetches.test.ts; unifies the API-constant-name reconstruction path.
