+++ src/client/App.svelte
@@ -137,32 +137,19 @@ wireActionDeps({
   openFindBar: () => {},
 });
 
-// Guard: only dispatch when visibility actually changes. Without this, every
-// call to updateSettings() (which replaces the entire settings object even
-// when showAuthorship is unchanged) would dispatch a transaction, causing
-// FormattingToolbar's tick++ listener to fire inside Svelte's effect flush
-// and eventually exceed the 1000-update depth limit (#613).
-//
-// First-run note: on the editor's initial availability we record the current
-// `visible` value WITHOUT dispatching. The authorship plugin reads
-// `localStorage[AUTHORSHIP_TOGGLE_KEY]` at construction time
-// (`useTandemSettings.svelte.ts` mirrors `showAuthorship` to that key on every
-// write) so the editor already starts in the correct state. Dispatching on
-// the first run was needed in the React era when the plugin had no localStorage
-// hook, and it is exactly the path that produced #613's prod-only effect-depth
-// loop — likely chained through y-prosemirror's transaction queue under prod's
-// tighter scheduling.
-let _lastAuthorshipVisible: boolean | undefined;
+// The authorship plugin reads its initial visibility from localStorage at
+// construction time, so dispatch only on subsequent changes — first-run
+// dispatch was the path that produced an effect-depth loop under prod
+// scheduling (transaction → tick → effect rerun → dispatch …).
+let lastDispatchedAuthorship: boolean | null = null;
 $effect(() => {
   const ed = editor;
   if (!ed) return;
   const visible = settingsState.settings.showAuthorship;
-  if (_lastAuthorshipVisible === undefined) {
-    _lastAuthorshipVisible = visible;
-    return;
-  }
-  if (_lastAuthorshipVisible === visible) return;
-  _lastAuthorshipVisible = visible;
+  if (lastDispatchedAuthorship === visible) return;
+  const firstRun = lastDispatchedAuthorship === null;
+  lastDispatchedAuthorship = visible;
+  if (firstRun) return;
   untrack(() => {
     const tr = ed.state.tr.setMeta(authorshipPluginKey, { type: "toggle", visible });
     ed.view.dispatch(tr);
@@ -194,12 +181,7 @@ let activeLeftRailTab = $state<"annotations" | "chat" | "outline">(
   settingsState.settings.leftRailTabs[0] ?? "annotations",
 );
 
-// Reconcile active-tab pointers whenever the persisted rail arrays change
-// (e.g. after settings load, cross-rail move, or external settings update).
-//
-// `untrack` around the write so the write back to activeLeftRailTab /
-// activeRailTab doesn't form a self-dep with this effect's includes() read.
-// Defensive against the prod-only #613 effect_update_depth shape.
+// `untrack` so the write doesn't form a self-dep with the includes() read.
 $effect(() => {
   const leftTabs = settingsState.settings.leftRailTabs;
   if (!leftTabs.includes(activeLeftRailTab)) {
