+++ src/server/mcp/file-opener.ts
@@ -195,10 +195,8 @@ export async function openFileFromContent(
   const id = docIdFromPath(syntheticPath);
 
   const doc = getOrCreateDocument(id);
-  // Route the upload through the shared batched populate so the upload path
-  // also fixes #609 (large markdown drag-drop freeze). Display name is the
-  // uploaded filename, not the synthetic upload path, so notifications don't
-  // leak the internal path shape to the user.
+  // Display name is the uploaded filename, not the synthetic upload path, so
+  // notifications don't leak the internal path shape to the user.
   await populateDocFromContent(doc, format, content, {
     displayName: fileName,
     dedupSource: syntheticPath,
@@ -403,15 +401,10 @@ interface PopulateContext {
 }
 
 /**
- * Load file content from disk into the Y.Doc. Thin wrapper around
- * populateDocFromContent — reads the buffer once (the caller has already
- * validated `resolved` via resolveAndValidatePath: size limit, extension
- * allowlist, UNC rejection) and delegates the parse + transact + cleanup logic
- * to the shared helper used by openFileFromContent.
+ * Load file content from disk into the Y.Doc. Reads the buffer once and
+ * delegates the parse + transact + cleanup to populateDocFromContent.
  */
 async function loadContentIntoDoc(doc: Y.Doc, format: string, resolved: string): Promise<void> {
-  // Single read site collapses what would otherwise be multiple fs.readFile
-  // call sites into one — fewer "uncontrolled path" sinks for CodeQL.
   const buffer = await fs.readFile(resolved);
   await populateDocFromContent(doc, format, buffer, {
     displayName: path.basename(resolved),
@@ -422,23 +415,12 @@ async function loadContentIntoDoc(doc: Y.Doc, format: string, resolved: string):
 /**
  * Shared populate path for openFileByPath (disk) and openFileFromContent
  * (upload). Async I/O and parsing happen OUTSIDE the transaction; the Y.Doc
- * mutation runs INSIDE one doc.transact(..., MCP_ORIGIN). Without batching,
- * mdastToYDoc fires thousands of tiny CRDT updates that flood y-prosemirror
- * and freeze the client on large markdown docs (#609). Precedent:
- * clearAndReload and reloadFromDisk (same file) use the same shape — async
- * I/O outside, Y.Doc mutation inside a single MCP_ORIGIN transact.
+ * mutation runs INSIDE one doc.transact(..., MCP_ORIGIN) so mdastToYDoc's
+ * many tiny inserts arrive as one update.
  *
  * MCP_ORIGIN is safe here (Critical Rule #2): the durable-annotation sync
  * observer and the channel event queue are attached later via
- * wireAnnotationStore (called from finalizeDocOpen, which runs after this
- * returns). No observer is watching during initial populate, so there is no
- * echo to suppress.
- *
- * On the success path, injectCommentsAsAnnotations' inner MCP_ORIGIN transact
- * is flattened by Yjs into the outer one — inner is a no-op wrapper. On the
- * failure path, the cleanup transact in the catch is a fresh top-level
- * transact: Yjs has already run the failed transact's finally
- * (afterTransaction + update emit) before the catch fires.
+ * wireAnnotationStore, after this returns — no observer to echo.
  */
 async function populateDocFromContent(
   doc: Y.Doc,
@@ -473,28 +455,22 @@ async function populateDocFromContent(
     applyToDoc = () => {
       htmlToYDoc(doc, html);
       if (comments.length > 0) {
-        // injectCommentsAsAnnotations opens an inner MCP_ORIGIN transact that
-        // Yjs flattens into ours, so writes land directly on this transaction's
-        // change set. Yjs does NOT roll back inner-transact writes on throw —
-        // without the snapshot/undo dance below, the doc would commit with N-of-M
-        // partial comments and re-open would hit importAnnotationId dedup,
-        // permanently dropping the failing comment with no surfaced error.
+        // Yjs does not roll back inner-transact writes on throw, so snapshot
+        // pre-inject annotation keys and undo any new ones if inject fails —
+        // otherwise an N-of-M commit collides with importAnnotationId dedup
+        // on next open and silently drops the failing comment.
         const annotMap = doc.getMap(Y_MAP_ANNOTATIONS);
         const before = new Set(annotMap.keys());
         try {
           injectCommentsAsAnnotations(doc, comments);
         } catch (err) {
-          for (const k of annotMap.keys()) {
+          for (const k of [...annotMap.keys()]) {
             if (!before.has(k)) annotMap.delete(k);
           }
           console.error(
             "[docx-comments] inject failed mid-transact; document loads without imported comments:",
             err,
           );
-          // Symmetric with the extract-failure notification ~30 lines above:
-          // both modes mean "docx loaded without (some) comments" from the
-          // user's perspective. Distinct dedupKey namespace so a file that
-          // hits both modes shows two toasts, not one collapsed one.
           pushNotification({
             id: generateNotificationId(),
             type: "annotation-error",
@@ -526,16 +502,10 @@ async function populateDocFromContent(
       doc.transact(() => {
         const fragment = doc.getXmlFragment("default");
         fragment.delete(0, fragment.length);
-        // Defensive: htmlToYDoc/loadMarkdown/populateYDoc shouldn't touch
-        // annotations, but injectCommentsAsAnnotations can leave partial entries
-        // even if the inner containment caught the throw (Yjs does not roll
-        // back inner-transact writes). REPLIES is not touched during populate,
-        // so leave it alone.
-        // MAINTENANCE: if a future populate helper writes to Y_MAP_ANNOTATION_REPLIES
-        // (e.g. importing a format with threaded comments), this cleanup must
-        // clear them too — the cleanup is silently incomplete until that happens.
+        // injectCommentsAsAnnotations can leave partial entries even when its
+        // own catch fires (Yjs does not roll back inner-transact writes).
         const annotations = doc.getMap(Y_MAP_ANNOTATIONS);
-        annotations.forEach((_, k) => annotations.delete(k));
+        for (const k of [...annotations.keys()]) annotations.delete(k);
       }, MCP_ORIGIN);
     } catch (cleanupErr) {
       console.error(
@@ -543,10 +513,8 @@ async function populateDocFromContent(
         cleanupErr,
       );
     }
-    // Keep the first console.error argument a static literal; pass
-    // user-controlled values (format, displayName, err) as trailing args so
-    // util.format never treats them as a format string. Defuses CodeQL's
-    // externally-controlled format string alert.
+    // Static-literal first arg; user-controlled values arrive as trailing args
+    // so util.format doesn't treat them as a format string.
     console.error(
       "[Tandem] populateDocFromContent: populate failed; partial state cleared before rethrow.",
       { format, displayName: ctx.displayName },
